refactor sveltia admin package
