+++ packages/sveltia-admin/__tests__/sveltia/inline-translation.spec.ts
@@ -28,8 +28,8 @@ vi.mock("virtual:lightnet/sveltiaAdminConfig", () => ({
   },
 }))
 
-import { mediaItemCollection } from "../../src/sveltia/collections/content/media-items"
-import { inlineTranslation } from "../../src/sveltia/utils/inline-translation"
+import { inlineTranslation } from "../../src/collections/fields/inline-translation"
+import { mediaItemCollection } from "../../src/collections/media-items"
 
 test("Should require only default locale in inline translation fields", () => {
   const field = inlineTranslation({ name: "label" }) as {
