+++ packages/sveltia-admin/src/Admin.astro
@@ -1,7 +1,7 @@
 ---
 import config from "virtual:lightnet/sveltiaAdminConfig"
 
-import { pathWithBase } from "./utils/paths"
+import { pathWithBase } from "./utils/path-with-base"
 ---
 
 <!doctype html>
