+++ packages/sveltia-admin/src/collections/media-items.ts
@@ -2,8 +2,8 @@ import type { Collection } from "@sveltia/cms"
 import config from "virtual:lightnet/config"
 import sveltiaAdminConfig from "virtual:lightnet/sveltiaAdminConfig"
 
-import { inlineTranslation } from "../../utils/inline-translation"
-import { projectPath } from "../../utils/path"
+import { projectPath } from "../utils/paths"
+import { inlineTranslation } from "./fields/inline-translation"
 import { languagesSelect } from "./languages"
 
 export const mediaItemCollection: Collection = {
